refactor: extract shared game detail components from desktop and web pages
The desktop game page (3,379 lines) and web share page (3,371 lines) were
94% identical. Extract all shared rendering into reusable components in
src/lib/game-detail/, reducing both pages to ~150-line thin wrappers.

Components: GameDetailView (orchestrator), EventsTab, LawsTab, TechsTab,
YieldsTab (data-driven via YIELD_CHART_CONFIG), MilitaryTab, CitiesTab,
ImprovementsTab, MapTab, SettingsTab, plus helpers.ts for shared types
and pure functions.

Context-specific UI uses Svelte 5 snippet props (headerActions for
desktop ShareControl, preTabs for web share banner). Map turn slider
uses optional onMapTurnChange callback.

Net reduction: 6,750 → 3,580 lines (47% less code, zero duplication).
