+++ src/lib/game-detail/index.ts
@@ -1,0 +1,1 @@
+export { default as GameDetailView } from "./GameDetailView.svelte";
