+++ web/src/lib/game-detail
@@ -1,0 +1,1 @@
+../../../src/lib/game-detail
